chore: rename to page service
